Title Screen and Pause menu activated.
NOTE: To enable pausing in a scene, go to the prefabs folder, and then drag both the Pause Canvas and GameManager prefabs into the scene.  Click on the GameManager and drag the Pause Canvas object in the scene into its place in the Inspector.  Finally, expand the Pause canvas, find the "Resume" button, and point it to the GameManager void "TogglePause".